+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - There now is a checkbox to allow to create or remove an autostart desktop entry
   (for [#6](https://github.com/pbek/nixbit/issues/6))
+  - The autostart entry can now be forced to be created with the external config file `/etc/nixbit.conf`
+  - There also is a new nix module option `services.nixbit.forceAutostart` for this
 - Fix warnings caused by the app icon SVG file (for [#2](https://github.com/pbek/nixbit/issues/2))
 
 ## 0.1.2